frame: use AdwTabView for frame contents
Instead of using GtkStack, we can use AdwTabView now that we depend on
libadwaita. This is useful so that we are able to use AdwTabBar for
primary content instead of recreating that widget.

As part of this, GtkStackSwitcher was imported and extended in the same
way we previously handled extra DnD however the page list comes from the
panel's AdwTabPage selection model instead.

A followup commit will add support for tab bars as an alternate
PanelFrameHeader implementation.