tabbar: add simple tabbar implementation using libadwaita
This will still need quite a bit of work related to when/how/what can be
closed (and if we need to move the panel back). Also, DnD between frames
and tabbar area needs work.

Additionally, DnD highlights don't work well with widgets that draw above
the drop location, so some work is needed there too.