tabbar: add close button to frame
If the frame is closeable, add a close frame button to the tabbar. Also,
always keep the menu visible and toggle close button instead, which was
really the point of the original menu (to be able to close the frame).

Additionally, use pan button for the frame menu, which is closer to what
we were doing before with Builder.