frame: make sure reveal properties are notified
I don't love how this is done, but we need a reliable way to ensure that
the dock get's the reveal properties set.

Personally, I wish I could just write all this over from scratch because
it's all sort of nightmarish now that I've seen the direction GTK 4 went
and how this all reminds me of GTK 3.