action-muxer: add support for class actions and property actions
This gives us most of what we had from GtkWidgetClass in a convenient
action muxer we can control. We don't get the part of action muxer that
lets us observe, but this at least gets us something we can group/insert
into GTK elsewhere.

Beyond that, we can probably make this useful outside of just
widgets so that we can have it in libide as well on IdeObject. Therefore
I tried to avoid the GTK'isms on the public API itself, even if they
are type compatible.