dock: rename PanelDockPosition to PanelArea
This has been a longstanding issue with the API that we need to fix before
we do a 1.0 and commit to some sort of semi-stability. Sorry for the late
cycle change, but given the amount of things that had to be done across
a number of modules, this is the best I could do.

Primarily, Builder is the consumer that drives the API, but after our 1.0,
we can look at how to better serve other applications as well.

Related #14