layout: implement PanelLayout and PanelLayoutItem
This creates a layout object which can be populated with items. The items
have an id, position, type-hint, and metadata. One of those items or more
must be set for the item to be of any use.

The goal here is for the PanelDock to use this structure to populate state
about the interface and save it to GVariant to be restored later.