types: add panel-types.h and use it
This allows us more flexibility with coordinating objects and how they
can reference each others types. It also tends to be a faster way of
dealing with header inclusions, but that's beside-the-point for now.

This uses the panel-types.h from headers, but also fixes the sources to
contain the include for the actual type definition, functions, etc.