save-dialog: implement save dialog visuals
This still doesn't pump the save delegates, but it does start to implement
the bits that are relevant for the mockups with how drafts are handled, as
well as 1 vs N documents.

I had to shove the list in a preferences page so that we could scroll in
case things get too large.

Additionally, the button visiblity is a total hack, and we probably want
proper API for that in libaddwaita directly.