frametabbar: disable ctrl+home/end variants by default
Currently we're mostly just used from Builder, and we don't have a way
to really plumb this API through. So do the right thing for the places
where we need it.

Eventually, we'll probably want to just expose the tab view as part of
the PanelFrame since AdwTabView is final.