menu-manager: add PanelMenuManager
This brings the menu manager from Builder which has the ability to merge
and unmerge menus at runtime.

It also provides some mutability which is useful when you need to update
accelerators due to user input.

The idea here is that this combines with other bits of Builder inside of
libpanel so that it can be used in other "largish" applications.